reducers changed to separate store better
