+++ client/src/components/presenters/userProfilePresenter.tsx
@@ -1,13 +1,13 @@
 import React, { useEffect, useState } from "react";
 import { connect } from "react-redux";
-import { orderMade } from "../../redux/actions/orders";
+import { orderPlaced } from "../../redux/actions/user";
 import UserProfile from "../views/userProfile";
 import { removeFavorite } from "../../redux/actions/user";
 
 export const UserProfilePresenter = ({
   orders,
   user,
-  orderMade,
+  orderPlaced,
   favorites,
   removeFavorite,
 }) => {
@@ -28,14 +28,14 @@ export const UserProfilePresenter = ({
 const mapStateToProps = (store) => {
   return {
     user: store.user.username,
-    orders: store.orders.userOrders,
+    orders: store.user.userOrders,
     favorites: store.user.favorites,
   };
 };
 
 const mapDispatchToProps = (dispatch) => {
   return {
-    orderMade: (beverage) => dispatch(orderMade(beverage)),
+    orderPlaced: (beverage) => dispatch(orderPlaced(beverage)),
     removeFavorite: (name) => dispatch(removeFavorite(name)),
   };
 };
